Bug 285 fix move set race condition.
During a move set it is possible on subscriber node, that isn't the
old or new origin, for remoteWorkerThread_$oldorigin to start
processing a SYNC event before the ACCEPT_SET is processed
by the remoteWorkerThread_$neworigin.  If the ACCEPT_SET
transaction then commits before the update sl_setsync
query is executed by the remoteWorkertThread_$oldorigin
the update will update a row for the new origin when
it intends to update the old origin.  This messes up
sl_setsync.

This change will add a condition to the WHERE clause so
a remoteWorkerThread won't UPDATE rows for a different
origin