bug 304 - record sequence values as part of a EXECUTE_SCRIPT
Prior to 2.2 the EXECUTE_SCRIPT command result in a DDL_SCRIPT event
and a SYNC event generated as part of the same transaction. That
SYNC event also replicates the value of any sequences to the replica
before the DDL SCRIPT command.

In 2.2 the DDL script is replicated as part of a normal SYNC
as rows in sl_log_script.  This commit stores any sequence changes
as extra elements in the cmdargs array.  These sequence values are
then set before the DDL_SCRIPT is executed on the replica

This commit includes a unit test change to reproduce this issue
the test now passes