bug 310 - only issue the notify Restart if we actually change configuration
The FAILOVER_NODE event is processed in two parts by slon.
First failedNode(..) is called which reconfigures sl_subscribe,
sl_path and then the listen network. Once this is done notify
is used to signal the slon to restart.

Then the slon commits this transactions and waits for any
pending events from the failed origin to arrive from a third
provider. Slon then continues with the failover (failoverSet_int).

The problem was that when slon restarts it repeats the processing
the FAILOVER_NODE. It was possible for slon to keep restarting
before it gets to be caught up and recording the FAILOVER_NODE
event as processed.

In this patch we only restart slon if the configuration actually
is changed. After the first restart the configuration shouldn't
need changing so there should not be anymore restarts.