bug 313 - don't error out in applying EXECUTE SCRIPT sequence updates
Since bug 304 was fixed we have been including the list of replicated
sequences and current values in the arguments on sl_log_script rows.
If some sequences are replicated to some nodes but not others then the
execute script will fail to apply on the nodes that don't have those
sequences.

This patch ignores missing sequences during an execute script
instead of erroring out.