Only include the libpq header directory OR the server header directory but not both
This addresses bug 315 where a debian install had libpq-dev 9.3 installed
and postgresql-server-dev 8.4 installed.  The libpq headers are in /usr/include/postgresql
but the server headers are in /usr/lib/postgresql/8.4/server.

configure checks were failing because we were pulling in a mixutre of headers
from both versions.  The only slony component that needs the server
headers are the backend functions and they shouldn't need to use libpq headers.

This patch sets CPPFLAGS as approriate for one or the other but not both