Fix for Bug 321
Bug 273 was fixed by no longer having the event provider
forced into the list of provider nodes. This caused bug
321 as a side effect slon is often adding then removing
(connecting and disconnecting) providers for nodes
that don't provide any real data.

A node that isn't an origin for a subscription still
might have rows in sl_log_script that need to be
processed.  We need to check some provider for
sl_log_script changes. Since the sl_listen network
for non-subscription providing nodes can have multiple
li_provider nodes for a given li_origin/li_receiver
we just make sure that ANY of those nodes is an estalished
provider.  We don't need it to be the event provider.

If the provider in the list is too far behind we abort
sync_event() pause and try again.  We also need to check
to see if the cluster configuration has been changed
by another worker during this time.