Remove HAVE_POSIX_SIGNALS from config.h
Compiling slony against PG 9.4 gives compiler warnings because the
definition of HAVE_POSIX_SIGNALS has changed.  This is defined in pg_config
so we do not really need to redefine it in the config.h slony provides.

We leave the configure check in place to we just don't include it in
the header