Bug 345
The strtok_r on the node list was adding replacing the comma
with a NULL causing only 1 node to get inserted into sl_event.
This mean any nodes that where cascaded only saw 1 of the nodes
in the drop node list.

Call strtok_r on a copy of the list so it can safely be munged.

Also add in a DropNode cluster test to demonstrate test this issue