Bug 354
The RebuildListenEntries for a failed node
that provides no subscriptions deletes all listen
entries from that node.  The remaining nodes
might still need to listen for these events
if the failover target has them.  Otherwise
the FAILOVER_NODE event will wait for this event
but it will never come.

Leave in the listen paths for failed nodes
that don't provide any sets