Explicitly disallow createEvent and data changes in the same transaction.
The same transaction should not have both a row in sl_event and log data
to sl_log_1 or sl_log_2 with the logTrigger.  This is not something slonik
would ever do, and the data changes that the logTrigger logs might not get
replicated as part of the event processing(ie for many non SYNC events).
it is also unclear what order the data should be applied in with respect to
the actions of the non-SYNC event.