Fix issue where DDL issued against a non-origin node can be duplicated.
The code intended to only use the provider that the
event was received from when querying sl_log_script but
not sets however the old queries in the provider structure
from previous syncs were not blanked and those queries were
being issued against the other providers.

Blank the provider query for providers we won't be using and
ignore those providers.